Accommodate rep log during mining cycle
Switch between using two reputation logs; one is able to be appended
to (this is the 'active' reputation log), and the other is frozen
until the current reputation mining cycle is over. Once the reputation
cycle is complete, the frozen cycle is deleted, which cycle is considered
the 'active' one is switched, and the process can repeat.